+++ src/Datepicker/types.ts
@@ -1,4 +1,5 @@
 import {Dayjs} from 'dayjs';
+import {ICommon} from "../typing";
 
 export interface ICurrentDayList {
     isActive: boolean
@@ -9,4 +10,14 @@ export interface ICurrentDayList {
     date: Dayjs,
     dayNumber: number,
     onClick: () => void
+}
+
+
+
+export interface IDatepickerProps extends ICommon{
+    value?: string;
+    onChange: (newDate: string) => void;
+    format?: string
+    onChangeYearMonthPanel?: (yearMonth: { year: number, month: number }) => void;
+    isVisibleSetToday?: boolean;
 }
