+++ src/index.ts
@@ -1,12 +1,12 @@
 import './styles.css';
 
 export {default as Datepicker} from './Datepicker';
-export {default as RangeDatepicker} from './RangeDatepicker';
-export {default as Timepicker} from './Timepicker';
-export {default as Timepicker2} from './Timepicker2';
-export {default as DateTimepicker} from './DateTimepicker';
-export {default as DateTimepicker2} from './DateTimepicker2';
-export {default as RangeTimeDatepicker} from './RangeTimeDatepicker';
+// export {default as RangeDatepicker} from './RangeDatepicker';
+// export {default as Timepicker} from './Timepicker';
+// export {default as Timepicker2} from './Timepicker2';
+// export {default as DateTimepicker} from './DateTimepicker';
+// export {default as DateTimepicker2} from './DateTimepicker2';
+// export {default as RangeTimeDatepicker} from './RangeTimeDatepicker';
 
 export {default as datepickerElClassNames} from './el-class-names';
-export * from './typing';
+export type {IRangeDateValue, IRangeDateTimeValue, EDateRange, EDateTimeRange, ITimeObj} from './typing';
